compares/harnesses/atheris/requirements.txt
# Atheris harness dependencies.
# Atheris is not Windows-compatible; run on Linux / WSL2 / macOS.
# Installing atheris triggers a native build step; Clang is required.
atheris==2.3.0
pysam==0.22.1
biopython==1.85
